Break out the OSD data holder from Common/System/System.h, into OSD.cpp/h
