+++ Core/Dialog/SavedataParam.cpp
@@ -22,7 +22,7 @@
 #include "Common/Data/Format/ZIMLoad.h"
 #include "Common/Serialize/Serializer.h"
 #include "Common/Serialize/SerializeFuncs.h"
-#include "Common/System/System.h"
+#include "Common/System/OSD.h"
 #include "Common/StringUtils.h"
 #include "Core/Config.h"
 #include "Core/Reporting.h"
