+++ Core/HLE/sceNetAdhoc.cpp
@@ -39,7 +39,7 @@
 #include "Common/Serialize/Serializer.h"
 #include "Common/Serialize/SerializeFuncs.h"
 #include "Common/Serialize/SerializeMap.h"
-#include "Common/System/System.h"
+#include "Common/System/OSD.h"
 #include "Common/Thread/ThreadUtil.h"
 #include "Common/TimeUtil.h"
