+++ GPU/Common/PresentationCommon.cpp
@@ -24,6 +24,7 @@
 
 #include "Common/System/Display.h"
 #include "Common/System/System.h"
+#include "Common/System/OSD.h"
 #include "Common/File/VFS/VFS.h"
 #include "Common/VR/PPSSPPVR.h"
 #include "Common/Log.h"
