+++ UI/NativeApp.cpp
@@ -61,6 +61,7 @@
 #include "Common/System/Display.h"
 #include "Common/System/Request.h"
 #include "Common/System/System.h"
+#include "Common/System/OSD.h"
 #include "Common/System/NativeApp.h"
 
 #include "Common/Data/Text/I18n.h"
