+++ UI/RetroAchievements.cpp
@@ -31,12 +31,13 @@
 #include "Common/File/Path.h"
 #include "Common/File/FileUtil.h"
 #include "Common/Net/HTTPClient.h"
+#include "Common/System/OSD.h"
+#include "Common/System/System.h"
 #include "Common/System/NativeApp.h"
 #include "Common/TimeUtil.h"
 #include "Common/Data/Text/I18n.h"
 #include "Common/Serialize/Serializer.h"
 #include "Common/StringUtils.h"
-#include "Common/System/System.h"
 #include "Common/Crypto/md5.h"
 #include "Common/UI/IconCache.h"
