+++ UWP/CommonUWP/CommonUWP.vcxproj
@@ -68,7 +68,7 @@
   <PropertyGroup>
     <GenerateManifest>false</GenerateManifest>
   </PropertyGroup>
-  <ItemDefinitionGroup  Condition="'$(Platform)' == 'Win32' or '$(Platform)' == 'x64'" Label="Configuration">
+  <ItemDefinitionGroup Condition="'$(Platform)' == 'Win32' or '$(Platform)' == 'x64'" Label="Configuration">
     <ClCompile>
       <CompileAsWinRT>false</CompileAsWinRT>
       <SDLCheck>false</SDLCheck>
@@ -214,6 +214,7 @@
     <ClInclude Include="..\..\Common\SysError.h" />
     <ClInclude Include="..\..\Common\System\Display.h" />
     <ClInclude Include="..\..\Common\System\NativeApp.h" />
+    <ClInclude Include="..\..\Common\System\OSD.h" />
     <ClInclude Include="..\..\Common\System\Request.h" />
     <ClInclude Include="..\..\Common\System\System.h" />
     <ClInclude Include="..\..\Common\Thread\Channel.h" />
@@ -352,6 +353,7 @@
     <ClCompile Include="..\..\Common\OSVersion.cpp" />
     <ClCompile Include="..\..\Common\StringUtils.cpp" />
     <ClCompile Include="..\..\Common\System\Display.cpp" />
+    <ClCompile Include="..\..\Common\System\OSD.cpp" />
     <ClCompile Include="..\..\Common\System\Request.cpp" />
     <ClCompile Include="..\..\Common\Thread\ThreadUtil.cpp" />
     <ClCompile Include="..\..\Common\Thread\ThreadManager.cpp" />
